doc: clarify installation with pip
The requirements file is not necessary for most users.  It is only
intended for advanced deployers interested in creating a specific
deployment.  The setup script already specifies the required packages,
and `pip` or `setuptools` can use the default information.

We still use the supplied `requirements.txt` file on GitLab.com's CI
runners.