tests: add reproducibility tests
Two reproducibility tests are added to the `tests` directory.

The first compares the result of the `withlogdet=False` compression
option, which is not recommended, on the real JLA data with their own
compression results in their paper (B14).  This should reproduce a
result close to JLA's own compression as measured by the KL divergence.

The other compares the results of the default, `withlogdet=True`
compression with the M16 paper's results (machine-readable tables A1 and
A2 from the paper).  This should reproduce a result that is element-wise
close to those in the M16 tables.

To run the reproducibility tests, the real JLA data must be present,
although they're not part of this source distribution.  On GitLab.com
they are stored in a private Docker image, and are available to the
automatic CI environment.

The tests can be run with pytest and are managed by tox.